+++ src/win32/RP_PropertyStore.cpp
@@ -129,6 +129,7 @@ const array<RP_PropertyStore_Private::MetaDataConv, static_cast<size_t>(Property
 	{&PKEY_RomProperties_PixelFormat, VT_BSTR},		// Pixel Format
 	{&PKEY_RomProperties_RegionCode, VT_BSTR},		// Region Code
 	{&PKEY_Category, VT_BSTR},				// Category
+	{&PKEY_RomProperties_RippedBy, VT_BSTR},		// Ripped By
 }};
 
 // Win32 SDK doesn't have this.